Re-factored all to isolate child/base reg classes Now the base class Regressor and MLP are set in a separated script. Similarly, all the child classes containing each individual regressor are separated into an isolated script which holds all hyperparams and getter methods to extract the corresponding values in reg_train.
